refactor(backup): split databaseBackupService into focused modules
Break the 1220-line databaseBackupService.ts into databaseBackup/:
constants (paths/pattern), types (merge summary + MERGEABLE_TABLES),
sqliteHelpers (generic merge primitives), tableMerges (per-table merge
logic + executeDatabaseMerge), and backupFiles (validate/list/create/
reinit + temp-import handling).

databaseBackupService.ts stays the public API facade: export/import/
preview/merge/getLastBackupInfo/restore/cleanup, re-exporting the
DatabaseMergeSummary type. Consumers use `import * as` and only touch
those exports, so nothing downstream changes. Behavior is unchanged;
the entry file drops 1220 -> 301 lines.
